pyproject.toml
[tool.poetry]
name = "silksong-announcer"
version = "0.1.0"
description = ""
authors = ["Silvan Schmidt <silvan.genius@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "silksong_announcer"}]

[tool.poetry.dependencies]
python = "^3.10"
python-steam-api = "^1.0.7"
beautifulsoup4 = "^4.11.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pyright = "^1.1.288"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
